requirements.txt
matplotlib~=3.5.1
numpy~=1.21.5
keras~=2.7.0
Pillow~=8.4.0
tensorflow~=2.7.0
scikit-learn~=1.0.1